+++ crates/entrypoint/src/cli/command/azure_devops/agent/mod.rs
@@ -1,3 +1,4 @@
 mod azure_devops_agent_cli;
 pub use azure_devops_agent_cli::*;
 pub mod package;
+pub mod pool;