+++ crates/entrypoint/src/cli/command/azure_devops/agent/pool/mod.rs
@@ -1,0 +1,3 @@
+mod azure_devops_agent_pool_cli;
+pub use azure_devops_agent_pool_cli::*;
+pub mod list;